+++ .gitignore
@@ -37,9 +37,10 @@ build/
 
 ### Application Properties ###
 src/main/resources/application-*.properties
-!src/main/resources/application.properties
+src/main/resources/application.properties
 src/main/resources/application-*.yml
 !src/main/resources/application.yml
+
 
 ### Logs ###
 logs/
