+++ src/main/java/com/HRManagementInformation/api/DepartmentController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.DepartmentSaveRequest;
-import com.HRManagementInformation.dto.request.DepartmentUpdateRequest;
+import com.HRManagementInformation.dto.request.department.DepartmentSaveRequest;
+import com.HRManagementInformation.dto.request.department.DepartmentUpdateRequest;
 import com.HRManagementInformation.dto.response.DepartmentResponse;
 import com.HRManagementInformation.entities.Department;
 import org.springframework.beans.factory.annotation.Autowired;
