+++ src/main/java/com/HRManagementInformation/api/DocumentTypeController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.DocumentTypeSaveRequest;
-import com.HRManagementInformation.dto.request.DocumentTypeUpdateRequest;
+import com.HRManagementInformation.dto.request.documentType.DocumentTypeSaveRequest;
+import com.HRManagementInformation.dto.request.documentType.DocumentTypeUpdateRequest;
 import com.HRManagementInformation.dto.response.DocumentTypeResponse;
 import com.HRManagementInformation.entities.DocumentType;
 import org.springframework.http.HttpStatus;
