+++ src/main/java/com/HRManagementInformation/api/EmployeeAttendanceController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.EmployeeAttendanceSaveRequest;
-import com.HRManagementInformation.dto.request.EmployeeAttendanceUpdateRequest;
+import com.HRManagementInformation.dto.request.employee.EmployeeAttendanceSaveRequest;
+import com.HRManagementInformation.dto.request.employee.EmployeeAttendanceUpdateRequest;
 import com.HRManagementInformation.dto.response.EmployeeAttendanceResponse;
 import com.HRManagementInformation.entities.EmployeeAttendance;
 import org.springframework.beans.factory.annotation.Autowired;
