+++ src/main/java/com/HRManagementInformation/api/EmployeeDocumentsController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.EmployeeDocumentsSaveRequest;
-import com.HRManagementInformation.dto.request.EmployeeDocumentsUpdateRequest;
+import com.HRManagementInformation.dto.request.employee.EmployeeDocumentsSaveRequest;
+import com.HRManagementInformation.dto.request.employee.EmployeeDocumentsUpdateRequest;
 import com.HRManagementInformation.dto.response.EmployeeDocumentsResponse;
 import com.HRManagementInformation.entities.EmployeeDocuments;
 import org.springframework.beans.factory.annotation.Autowired;
