+++ src/main/java/com/HRManagementInformation/api/InterviewController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.InterviewSaveRequest;
-import com.HRManagementInformation.dto.request.InterviewUpdateRequest;
+import com.HRManagementInformation.dto.request.interview.InterviewSaveRequest;
+import com.HRManagementInformation.dto.request.interview.InterviewUpdateRequest;
 import com.HRManagementInformation.dto.response.InterviewResponse;
 import com.HRManagementInformation.entities.Interviews;
 import org.springframework.beans.factory.annotation.Autowired;
