+++ src/main/java/com/HRManagementInformation/api/JobOfferController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.JobOfferSaveRequest;
-import com.HRManagementInformation.dto.request.JobOfferUpdateRequest;
+import com.HRManagementInformation.dto.request.jobOffer.JobOfferSaveRequest;
+import com.HRManagementInformation.dto.request.jobOffer.JobOfferUpdateRequest;
 import com.HRManagementInformation.dto.response.JobOfferResponse;
 import com.HRManagementInformation.entities.JobOffers;
 import org.springframework.beans.factory.annotation.Autowired;
