+++ src/main/java/com/HRManagementInformation/api/LeaveRequestController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.LeaveRequestSaveRequest;
-import com.HRManagementInformation.dto.request.LeaveRequestUpdateRequest;
+import com.HRManagementInformation.dto.request.leave.LeaveRequestSaveRequest;
+import com.HRManagementInformation.dto.request.leave.LeaveRequestUpdateRequest;
 import com.HRManagementInformation.dto.response.LeaveRequestResponse;
 import com.HRManagementInformation.entities.LeaveRequest;
 import org.springframework.http.HttpStatus;
