+++ src/main/java/com/HRManagementInformation/api/LeaveTypeController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.LeaveTypeSaveRequest;
-import com.HRManagementInformation.dto.request.LeaveTypeUpdateRequest;
+import com.HRManagementInformation.dto.request.leaveType.LeaveTypeSaveRequest;
+import com.HRManagementInformation.dto.request.leaveType.LeaveTypeUpdateRequest;
 import com.HRManagementInformation.dto.response.LeaveTypeResponse;
 import com.HRManagementInformation.entities.LeaveType;
 import org.springframework.http.HttpStatus;
