+++ src/main/java/com/HRManagementInformation/api/MonthlyGoalsController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.MonthlyGoalsSaveRequest;
-import com.HRManagementInformation.dto.request.MonthlyGoalsUpdateRequest;
+import com.HRManagementInformation.dto.request.monthlyGoals.MonthlyGoalsSaveRequest;
+import com.HRManagementInformation.dto.request.monthlyGoals.MonthlyGoalsUpdateRequest;
 import com.HRManagementInformation.dto.response.MonthlyGoalsResponse;
 import com.HRManagementInformation.entities.MonthlyGoals;
 import org.springframework.beans.factory.annotation.Autowired;
