+++ src/main/java/com/HRManagementInformation/api/RecruitmentController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.RecruitmentSaveRequest;
-import com.HRManagementInformation.dto.request.RecruitmentUpdateRequest;
+import com.HRManagementInformation.dto.request.recruitment.RecruitmentSaveRequest;
+import com.HRManagementInformation.dto.request.recruitment.RecruitmentUpdateRequest;
 import com.HRManagementInformation.dto.response.RecruitmentResponse;
 import com.HRManagementInformation.entities.Recruitment;
 import org.springframework.beans.factory.annotation.Autowired;
