+++ src/main/java/com/HRManagementInformation/api/UserController.java
@@ -5,8 +5,8 @@ import com.HRManagementInformation.core.config.modelMapper.IModelMapperService;
 import com.HRManagementInformation.core.result.Result;
 import com.HRManagementInformation.core.result.ResultData;
 import com.HRManagementInformation.core.utilies.ResultHelper;
-import com.HRManagementInformation.dto.request.UserSaveRequest;
-import com.HRManagementInformation.dto.request.UserUpdateRequest;
+import com.HRManagementInformation.dto.request.user.UserSaveRequest;
+import com.HRManagementInformation.dto.request.user.UserUpdateRequest;
 import com.HRManagementInformation.dto.response.UserResponse;
 import com.HRManagementInformation.entities.User;
 import org.springframework.http.HttpStatus;
