+++ src/main/java/com/HRManagementInformation/dto/request/department/DepartmentUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.HRManagementInformation.dto.request;
+package com.HRManagementInformation.dto.request.department;
 
 public class DepartmentUpdateRequest {
     private int id;
