+++ src/main/java/com/HRManagementInformation/dto/request/employee/EmployeeDocumentsSaveRequest.java
@@ -1,23 +1,14 @@
-package com.HRManagementInformation.dto.request;
+package com.HRManagementInformation.dto.request.employee;
 
 import java.sql.Date;
 
-public class EmployeeDocumentsUpdateRequest {
-    private int documentId;
+public class EmployeeDocumentsSaveRequest {
     private int userId;
     private String filePath;
     private Date uploadDate;
     private Date validUntil;
 
-    public EmployeeDocumentsUpdateRequest() {
-    }
-
-    public int getDocumentId() {
-        return documentId;
-    }
-
-    public void setDocumentId(int documentId) {
-        this.documentId = documentId;
+    public EmployeeDocumentsSaveRequest() {
     }
 
     public int getUserId() {
