+++ src/main/java/com/HRManagementInformation/dto/request/jobOffer/JobOfferSaveRequest.java
@@ -1,11 +1,18 @@
-package com.HRManagementInformation.dto.request;
+package com.HRManagementInformation.dto.request.jobOffer;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 import java.sql.Date;
 
-public class JobOfferUpdateRequest {
+public class JobOfferSaveRequest {
+    @Positive(message = "id pozitif bir değer olmak zorunda")
+    @NotNull(message = "id değeri null veya boş olamaz")
+    private int offerId;
+
+    @NotNull(message = "recruitment id değeri null veya boş olamaz")
+    private int recruitmentId;
+
     @NotNull(message = "teklif tarihi boş olamaz")
     private Date offerDate;
 
@@ -18,14 +25,36 @@ public class JobOfferUpdateRequest {
 
     private String feedback;
 
-    public JobOfferUpdateRequest() {
+    public JobOfferSaveRequest() {
     }
 
-    public JobOfferUpdateRequest(Date offerDate, double offeredSalary, String offerStatus, String feedback) {
+    public JobOfferSaveRequest(int offerId, int recruitmentId, Date offerDate, double offeredSalary, String offerStatus, String feedback) {
+        this.offerId = offerId;
+        this.recruitmentId = recruitmentId;
         this.offerDate = offerDate;
         this.offeredSalary = offeredSalary;
         this.offerStatus = offerStatus;
         this.feedback = feedback;
+    }
+
+    @Positive(message = "id pozitif bir değer olmak zorunda")
+    @NotNull(message = "id değeri null veya boş olamaz")
+    public int getOfferId() {
+        return offerId;
+    }
+
+    public void setOfferId(@Positive(message = "id pozitif bir değer olmak zorunda")
+                          @NotNull(message = "id değeri null veya boş olamaz") int offerId) {
+        this.offerId = offerId;
+    }
+
+    @NotNull(message = "recruitment id değeri null veya boş olamaz")
+    public int getRecruitmentId() {
+        return recruitmentId;
+    }
+
+    public void setRecruitmentId(@NotNull(message = "recruitment id değeri null veya boş olamaz") int recruitmentId) {
+        this.recruitmentId = recruitmentId;
     }
 
     @NotNull(message = "teklif tarihi boş olamaz")
